w6 d1

+++ Spring/WS/addressbook-portal/src/main/java/com/cts/addressbook/repo/ContactRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cts.addressbook.entity.Contact;
 
-public interface ContactRepo extends JpaRepository<Contact,Long> {
+public interface ContactRepo extends JpaRepository<Contact,Integer> {
 	
 }
